promote use of pkgin for binary packages
Instead of installing pkgin through the provisioning script, use
the related option in sysinst.

This has the advantage of making more obvious how to add additional
packages, as well as making better use of the networking setup by
the installer.

While at it, remove an unnecessary (and probably conflicting) call
to setup the network that is not needed to fetch and execute the
post_install script.